+++ fbpcs/CHANGELOG.md
@@ -17,7 +17,7 @@ Types of changes
 ### Added
 
 ### Changed
-
+  - Mark validate_container_definition as deprecated in PCSContainerService since it is no longer a public method in ContainerService in fbpcp
 ### Removed
 
 ## [1.10.0] - 2022-08-12